Update README.md
According to https://www.roboleary.net/vscode/2020/08/05/dont-need-extensions.html, auto renaming HTML/XML tags has been a native feature for more than a year, the default was just false. And for auto closing tags, it was also a native feature for more than a year, and the default has been true so no need to mention it.